requirements.txt
openai
faiss-cpu
langchain
pypdf
tqdm
streamlit
numpy
pandas
python-dotenv
langchain-openai
streamlit-chat
